pyproject.toml
[project]
name = "ProjectDocs"
version = "0.1.0"
description = "The documentations of the projects practicing which use for improve my CS self learning."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "babel==2.17.0",
    "backrefs==5.8",
    "beautifulsoup4==4.13.4",
    "bleach==6.2.0",
    "build==1.3.0",
    "cachecontrol==0.14.0",
    "cairocffi==1.7.1",
    "cairosvg==2.8.2",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "colorama==0.4.6",
    "colorlog==6.9.0",
    "comm==0.2.3",
    "contourpy==1.3.3",
    "coverage==7.10.6",
    "csscompressor==0.9.5",
    "cssselect2==0.8.0",
    "cycler==0.12.1",
    "debugpy==1.8.16",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "executing==2.2.1",
    "fastjsonschema==2.21.2",
    "filelock==3.15.4",
    "fonttools==4.59.2",
    "frontmatter==3.0.8",
    "ghp-import==2.1.0",
    "gitdb==4.0.10",
    "gitpython==3.1.44",
    "griffe==1.9.0",
    "htmlmin2==0.1.13",
    "idna==3.10",
    "iniconfig==2.1.0",
    "ipykernel==6.30.1",
    "ipython==9.5.0",
    "ipython-pygments-lexers==1.1.1",
    "ipywidgets==8.1.7",
    "jedi==0.19.2",
    "jieba==0.42.1",
    "jinja2==3.1.6",
    "joblib>=1.5.2",
    "jsmin==3.0.1",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.4.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "jupyterlab-pygments==0.3.0",
    "jupyterlab-widgets==3.0.15",
    "jupytext==1.17.3",
    "kiwisolver==1.4.9",
    "lxml==6.0.0",
    "markdown==3.8",
    "markdown-callouts==0.4.0",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.2",
    "matplotlib==3.10.6",
    "matplotlib-inline==0.1.7",
    "mdit-py-plugins==0.5.0",
    "mdurl==0.1.2",
    "mergedeep==1.3.4",
    "mistune==3.1.4",
    "mkdocs==1.6.1",
    "mkdocs-autorefs==1.4.2",
    "mkdocs-drawio==1.11.2",
    "mkdocs-get-deps==0.2.0",
    "mkdocs-git-revision-date-localized-plugin==1.4.7",
    "mkdocs-jupyter==0.25.1",
    "mkdocs-material==9.6.14",
    "mkdocs-material-extensions==1.3.1",
    "mkdocs-minify-plugin==0.7.1",
    "mkdocs-note==1.0.3",
    "mkdocs-obsidian-interactive-graph-plugin==0.3.0",
    "mkdocs-redirects==1.2.1",
    "mkdocs-rss-plugin==1.17.3",
    "mkdocs-static-i18n==1.3.0",
    "mkdocstrings==0.30.0",
    "mkdocstrings-python==1.16.12",
    "msgpack==1.0.8",
    "nbclient==0.10.2",
    "nbconvert==7.16.6",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "numpy==2.3.3",
    "packaging==25.0",
    "paginate==0.5.7",
    "pandas==2.3.2",
    "pandocfilters==1.5.1",
    "parso==0.8.5",
    "pathspec==0.12.1",
    "pexpect==4.9.0",
    "pillow==10.4.0",
    "platformdirs==4.3.8",
    "pluggy==1.6.0",
    "prompt-toolkit==3.0.52",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pycparser==2.22",
    "pygments==2.19.1",
    "pymdown-extensions==10.15",
    "pyparsing==3.2.4",
    "pypinyin==0.51.0",
    "pyproject-hooks==1.2.0",
    "pytest==8.4.2",
    "pytest-cov==6.2.1",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyyaml==5.1",
    "pyyaml-env-tag==1.1",
    "pyzmq==27.0.2",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "rpds-py==0.27.1",
    "scikit-learn>=1.7.2",
    "sgmllib3k==1.0.0",
    "six==1.17.0",
    "smmap==5.0.0",
    "soupsieve==2.7",
    "stack-data==0.6.3",
    "tinify==1.6.0",
    "tinycss2==1.4.0",
    "tornado==6.5.2",
    "traitlets==5.14.3",
    "typing-extensions==4.14.1",
    "tzdata==2024.1",
    "urllib3==2.4.0",
    "watchdog==6.0.0",
    "wcwidth==0.2.13",
    "webencodings==0.5.1",
    "widgetsnbextension==4.0.14",
]
